+++ FRCore/FRCore/Security/SymbolicLinkDetector.swift
@@ -16,14 +16,23 @@ public class SymbolicLinkDetector: JailbreakDetector {
     /// Initializes SymbolicLinkDetector instance
     public init() { }
     
-    /// Analyzes whether "/Applications", "/Library/Ringtones", "/Library/Wallpaper" directory is a symbolic link or not
+    /// Analyzes whether certain directories are symbolic links or not
     ///
-    /// - NOTE: As part of Jailbreak process, it is commonly known that Jailbreak process will overwrite the partition, and changes some directories as symbolic link as original file/directory should remain as it was. "/Applications", "/Library/Ringtones", and/or "/Library/Wallpaper" directory is one of them, and the class is checking whether "/Applications", "/Library/Ringtones", "/Library/Wallpaper" directory is found as symbolic link or not
+    /// - NOTE: As part of Jailbreak process, it is commonly known that Jailbreak process will overwrite the partition, and changes some directories as symbolic link as original file/directory should remain as it was
     ///
-    /// - Returns: returns 1.0 when "/Applications", "/Library/Ringtones", "/Library/Wallpaper" directory is found as a symbolic link; otherwise returns 0.0
+    /// - Returns: returns 1.0 when certain directories are found as symbolic links; otherwise returns 0.0
     public func analyze() -> Double {
         Log.v("\(self) analyzing")
-        let urls: [String] = ["/Applications", "/Library/Ringtones", "/Library/Wallpaper"]
+        let urls: [String] = [
+            "/var/lib/undecimus/apt", 
+            "/Applications",
+            "/Library/Ringtones",
+            "/Library/Wallpaper",
+            "/usr/arm-apple-darwin9",
+            "/usr/include",
+            "/usr/libexec",
+            "/usr/share"
+        ]
         
         for urlString in urls {
             let url = URL(fileURLWithPath: urlString)